requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-google-genai==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.18
pandas==2.1.4
numpy==1.24.3
streamlit-option-menu==0.3.6
plotly==5.17.0
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0